:recycle: split SafeLinearArith and SafeDivision into two files
