+++ grisette.cabal
@@ -70,7 +70,8 @@ library
       Grisette.Core.Data.Class.LogicalOp
       Grisette.Core.Data.Class.Mergeable
       Grisette.Core.Data.Class.ModelOps
-      Grisette.Core.Data.Class.SafeArith
+      Grisette.Core.Data.Class.SafeDivision
+      Grisette.Core.Data.Class.SafeLinearArith
       Grisette.Core.Data.Class.SEq
       Grisette.Core.Data.Class.SignConversion
       Grisette.Core.Data.Class.SimpleMergeable
